+++ tests/unit/test_progress.py
@@ -14,10 +14,14 @@ class TestProgressAware(TestCase):
         function = progress_aware(lambda: None)
 
         function()
+        # first - called once for progress.clearln()
         mock_print.assert_called_once()
 
         progress.update_target("testing", 0, 100)
         function()
+        # second - progress.clearln() on refresh on update_target
+        # third - progress.print on refresh on update_target
+        # fourth - progress.clearln()
         self.assertEqual(4, mock_print.call_count)
 
         progress.finish_target("testing")
